Complete de the score and views route and refromating the Main files

+++ src/main/scala/org/esgi/project/models/Likes.scala
@@ -4,7 +4,7 @@ import play.api.libs.json.Json
 
 case class Likes(
                     _id: Int,
-                    score: Long
+                    score: Double
                  )
 object Likes {
     implicit val format = Json.format[Likes]
